requirements.txt
PyMuPDF==1.22.5
sentence-transformers==2.2.2
numpy==1.24.3
nltk==3.8.1
torch>=2.0.0
tqdm>=4.65.0
chromadb==0.4.22
python-magic>=0.4.27
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0
setuptools>=65.5.1
wheel>=0.38.4
cmake>=3.25.0 